pyproject.toml
[project]
name = "opendatajounalism"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "folium>=0.19.6",
    "geopandas>=1.0.1",
    "jageocoder>=2.1.8",
    "japanize-matplotlib>=1.1.3",
    "matplotlib>=3.10.3",
    "numpy>=1.26.4",
    "pandas>=1.5.3",
    "pandas-estat>=0.3.4",
    "plotly>=6.1.0",
    "pyproj>=3.7.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "shapely>=2.1.1",
]

[dependency-groups]
dev = [
    "bandit>=1.8.5",
    "black>=25.1.0",
    "mypy>=1.16.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.12.0",
    "safety>=3.5.2",
]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [".git", ".ruff_cache", ".venv", ".vscode"]

[tool.ruff.lint]
preview = true
select = [
    "ANN",  # type annotation
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "RUF",  # ruff specific rules
    "W",    # pycodestyle warnings
]
ignore = [
    "F401",  # unused import
    "F841",  # unused variable
]
unfixable = ["F401", "F841"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["ANN"]

[tool.ruff.lint.pydocstyle]
convention = "google"
